requirements.txt
# LangChain
langchain
langchain-community
langchain-postgres
langchain-google-genai

# Database
psycopg[binary]

# LLM & Embedding (Gemini)
google-genai

# API
fastapi
uvicorn

# Utils
python-dotenv

# Testing
pytest
httpx